services/ai-mcp-service/requirements.txt
fastapi
uvicorn
mcp
chromadb
openai
google-generativeai
pandas
deltalake
python-dotenv
pyarrow
